requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via fsspec
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via httpx
attrs==25.4.0
    # via aiohttp
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
datasets==4.4.2
    # via mycelium (pyproject.toml)
dill==0.4.0
    # via
    #   datasets
    #   multiprocess
filelock==3.20.3
    # via
    #   datasets
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.10.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via httpcore
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   mycelium (pyproject.toml)
    #   datasets
huggingface-hub==0.36.0
    # via
    #   datasets
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via torch
joblib==1.5.3
    # via scikit-learn
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.18
    # via datasets
networkx==3.6.1
    # via torch
numpy==2.4.1
    # via
    #   mycelium (pyproject.toml)
    #   datasets
    #   pandas
    #   scikit-learn
    #   scipy
    #   transformers
packaging==25.0
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
pandas==2.3.3
    # via datasets
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pyarrow==22.0.0
    # via datasets
pydantic==2.12.5
    # via mycelium (pyproject.toml)
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via mycelium (pyproject.toml)
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
regex==2025.11.3
    # via transformers
requests==2.32.5
    # via
    #   datasets
    #   huggingface-hub
    #   transformers
safetensors==0.7.0
    # via transformers
scikit-learn==1.8.0
    # via sentence-transformers
scipy==1.16.3
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.2.0
    # via mycelium (pyproject.toml)
six==1.17.0
    # via python-dateutil
sympy==1.14.0
    # via
    #   mycelium (pyproject.toml)
    #   torch
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.2
    # via transformers
torch==2.9.1
    # via sentence-transformers
tqdm==4.67.1
    # via
    #   datasets
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.57.3
    # via sentence-transformers
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   sentence-transformers
    #   torch
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
xxhash==3.6.0
    # via datasets
yarl==1.22.0
    # via aiohttp
